UIE-200 Ajax cleanup pt1  - moved Workspaces ajax subarea out of ajax.js and into a seperate ts module.  - converted ajax.js to ts.  - minimal Typescript conversion.  Added typed for data call returns, etc. will need to come from feature team folks.
