+++ src/libs/ajax/workspaces/Workspaces.ts
@@ -2,31 +2,9 @@ import { jsonBody } from '@terra-ui-packages/data-client-core';
 import _ from 'lodash/fp';
 import * as qs from 'qs';
 import { authOpts } from 'src/auth/auth-session';
-import { fetchAgora, fetchDrsHub, fetchGoogleForms, fetchOrchestration, fetchRawls } from 'src/libs/ajax/ajax-common';
-import { AzureStorage } from 'src/libs/ajax/AzureStorage';
-import { Billing } from 'src/libs/ajax/Billing';
-import { Catalog } from 'src/libs/ajax/Catalog';
-import { DataRepo } from 'src/libs/ajax/DataRepo';
-import { Dockstore } from 'src/libs/ajax/Dockstore';
-import { ExternalCredentials } from 'src/libs/ajax/ExternalCredentials';
-import { appIdentifier, fetchOk } from 'src/libs/ajax/fetch/fetch-core';
+import { fetchOrchestration, fetchRawls } from 'src/libs/ajax/ajax-common';
+import { fetchOk } from 'src/libs/ajax/fetch/fetch-core';
 import { GoogleStorage } from 'src/libs/ajax/GoogleStorage';
-import { Groups } from 'src/libs/ajax/Groups';
-import { Apps } from 'src/libs/ajax/leonardo/Apps';
-import { Disks } from 'src/libs/ajax/leonardo/Disks';
-import { Runtimes } from 'src/libs/ajax/leonardo/Runtimes';
-import { Metrics } from 'src/libs/ajax/Metrics';
-import { OAuth2 } from 'src/libs/ajax/OAuth2';
-import { SamResources } from 'src/libs/ajax/SamResources';
-import { Support } from 'src/libs/ajax/Support';
-import { TermsOfService } from 'src/libs/ajax/TermsOfService';
-import { User } from 'src/libs/ajax/User';
-import { Cbas } from 'src/libs/ajax/workflows-app/Cbas';
-import { CromwellApp } from 'src/libs/ajax/workflows-app/CromwellApp';
-import { WorkflowScript } from 'src/libs/ajax/workflows-app/WorkflowScript';
-import { WorkspaceData } from 'src/libs/ajax/WorkspaceDataService';
-import { WorkspaceManagerResources } from 'src/libs/ajax/WorkspaceManagerResources';
-import { getConfig } from 'src/libs/config';
 import { getTerraUser } from 'src/libs/state';
 import * as Utils from 'src/libs/utils';
 
@@ -55,37 +33,13 @@ const attributesUpdateOps = _.flow(
       : [{ op: 'AddUpdateAttribute', attributeName: k, addUpdateAttribute: v }];
   })
 );
-
-const CromIAM = (signal) => ({
-  callCacheDiff: async (thisWorkflow, thatWorkflow) => {
-    const { workflowId: thisWorkflowId, callFqn: thisCallFqn, index: thisIndex } = thisWorkflow;
-    const { workflowId: thatWorkflowId, callFqn: thatCallFqn, index: thatIndex } = thatWorkflow;
-
-    const params = {
-      workflowA: thisWorkflowId,
-      callA: thisCallFqn,
-      indexA: thisIndex !== -1 ? thisIndex : undefined,
-      workflowB: thatWorkflowId,
-      callB: thatCallFqn,
-      indexB: thatIndex !== -1 ? thatIndex : undefined,
-    };
-    const res = await fetchOrchestration(`api/workflows/v1/callcaching/diff?${qs.stringify(params)}`, _.merge(authOpts(), { signal }));
-    return res.json();
-  },
-
-  workflowMetadata: async (workflowId, includeKey, excludeKey) => {
-    const res = await fetchOrchestration(
-      `api/workflows/v1/${workflowId}/metadata?${qs.stringify({ includeKey, excludeKey }, { arrayFormat: 'repeat' })}`,
-      _.merge(authOpts(), { signal })
-    );
-    return res.json();
-  },
-});
-
-const Workspaces = (signal) => ({
+export const Workspaces = (signal) => ({
   list: async (fields, stringAttributeMaxLength) => {
     const lenParam = _.isNil(stringAttributeMaxLength) ? '' : `stringAttributeMaxLength=${stringAttributeMaxLength}&`;
-    const res = await fetchRawls(`workspaces?${lenParam}${qs.stringify({ fields }, { arrayFormat: 'comma' })}`, _.merge(authOpts(), { signal }));
+    const res = await fetchRawls(
+      `workspaces?${lenParam}${qs.stringify({ fields }, { arrayFormat: 'comma' })}`,
+      _.merge(authOpts(), { signal })
+    );
     return res.json();
   },
 
@@ -100,7 +54,7 @@ const Workspaces = (signal) => ({
   },
 
   getTags: async (tag, limit) => {
-    const params = { q: tag };
+    const params: any = { q: tag };
     if (limit) {
       params.limit = limit;
     }
@@ -108,11 +62,7 @@ const Workspaces = (signal) => ({
     return res.json();
   },
 
-  /**
-   * @param {string} workspaceId
-   * @param {string[]} [fields]
-   */
-  getById: async (workspaceId, fields = undefined) => {
+  getById: async (workspaceId: string, fields: string[] | undefined = undefined) => {
     let url = `workspaces/id/${workspaceId}`;
     if (fields) {
       url += `?${qs.stringify({ fields }, { arrayFormat: 'comma' })}`;
@@ -126,7 +76,10 @@ const Workspaces = (signal) => ({
 
     return {
       clone: async (body) => {
-        const res = await fetchRawls(`${root}/clone`, _.mergeAll([authOpts(), jsonBody(body), { signal, method: 'POST' }]));
+        const res = await fetchRawls(
+          `${root}/clone`,
+          _.mergeAll([authOpts(), jsonBody(body), { signal, method: 'POST' }])
+        );
         return res.json();
       },
 
@@ -165,7 +118,10 @@ const Workspaces = (signal) => ({
   },
 
   startBatchBucketMigration: async (body) => {
-    const response = await fetchRawls('workspaces/v2/bucketMigration', _.mergeAll([authOpts(), jsonBody(body), { signal, method: 'POST' }]));
+    const response = await fetchRawls(
+      'workspaces/v2/bucketMigration',
+      _.mergeAll([authOpts(), jsonBody(body), { signal, method: 'POST' }])
+    );
     return response.json();
   },
 
@@ -183,7 +139,10 @@ const Workspaces = (signal) => ({
       checkBucketLocation: GoogleStorage(signal).checkBucketLocation,
 
       details: async (fields) => {
-        const res = await fetchRawls(`${root}?${qs.stringify({ fields }, { arrayFormat: 'comma' })}`, _.merge(authOpts(), { signal }));
+        const res = await fetchRawls(
+          `${root}?${qs.stringify({ fields }, { arrayFormat: 'comma' })}`,
+          _.merge(authOpts(), { signal })
+        );
         return res.json();
       },
 
@@ -236,7 +195,10 @@ const Workspaces = (signal) => ({
               source: { namespace: configNamespace, name: configName, workspaceName: { namespace, name } },
               destination: { namespace: destConfigNamespace, name: destConfigName, workspaceName },
             };
-            const res = await fetchRawls('methodconfigs/copy', _.mergeAll([authOpts(), jsonBody(payload), { signal, method: 'POST' }]));
+            const res = await fetchRawls(
+              'methodconfigs/copy',
+              _.mergeAll([authOpts(), jsonBody(payload), { signal, method: 'POST' }])
+            );
             return res.json();
           },
 
@@ -273,10 +235,17 @@ const Workspaces = (signal) => ({
       },
 
       listSnapshots: async (limit, offset) => {
-        const res = await fetchRawls(`${root}/snapshots/v2?offset=${offset}&limit=${limit}`, _.merge(authOpts(), { signal }));
+        const res = await fetchRawls(
+          `${root}/snapshots/v2?offset=${offset}&limit=${limit}`,
+          _.merge(authOpts(), { signal })
+        );
         // The list snapshots endpoint returns a "snapshot" field that should really be named "snapshotId". Ideally, this should be fixed in the
         // backend, but we've sequestered it here for now.
-        return _.update('gcpDataRepoSnapshots', _.map(_.update('attributes', (a) => ({ ...a, snapshotId: a.snapshot }))), await res.json());
+        return _.update(
+          'gcpDataRepoSnapshots',
+          _.map(_.update('attributes', (a) => ({ ...a, snapshotId: a.snapshot }))),
+          await res.json()
+        );
       },
 
       snapshot: (snapshotId) => {
@@ -289,7 +258,10 @@ const Workspaces = (signal) => ({
           },
 
           update: (updateInfo) => {
-            return fetchRawls(snapshotPath, _.mergeAll([authOpts(), jsonBody(updateInfo), { signal, method: 'PATCH' }]));
+            return fetchRawls(
+              snapshotPath,
+              _.mergeAll([authOpts(), jsonBody(updateInfo), { signal, method: 'PATCH' }])
+            );
           },
 
           delete: () => {
@@ -340,7 +312,10 @@ const Workspaces = (signal) => ({
                 return res.json();
               },
 
-              outputs: () => fetchRawls(`${submissionPath}/workflows/${workflowId}/outputs`, _.merge(authOpts(), { signal })).then((r) => r.json()),
+              outputs: () =>
+                fetchRawls(`${submissionPath}/workflows/${workflowId}/outputs`, _.merge(authOpts(), { signal })).then(
+                  (r) => r.json()
+                ),
             };
           },
         };
@@ -370,7 +345,10 @@ const Workspaces = (signal) => ({
       },
 
       createEntity: async (payload) => {
-        const res = await fetchRawls(`${root}/entities`, _.mergeAll([authOpts(), jsonBody(payload), { signal, method: 'POST' }]));
+        const res = await fetchRawls(
+          `${root}/entities`,
+          _.mergeAll([authOpts(), jsonBody(payload), { signal, method: 'POST' }])
+        );
         return res.json();
       },
 
@@ -384,7 +362,10 @@ const Workspaces = (signal) => ({
       renameEntityType: (oldName, newName) => {
         const payload = { newName };
 
-        return fetchRawls(`${root}/entityTypes/${oldName}`, _.mergeAll([authOpts(), jsonBody(payload), { signal, method: 'PATCH' }]));
+        return fetchRawls(
+          `${root}/entityTypes/${oldName}`,
+          _.mergeAll([authOpts(), jsonBody(payload), { signal, method: 'PATCH' }])
+        );
       },
 
       deleteEntitiesOfType: (entityType) => {
@@ -408,11 +389,17 @@ const Workspaces = (signal) => ({
           entityNames
         );
 
-        return fetchRawls(`${root}/entities/batchUpsert`, _.mergeAll([authOpts(), jsonBody(body), { signal, method: 'POST' }]));
+        return fetchRawls(
+          `${root}/entities/batchUpsert`,
+          _.mergeAll([authOpts(), jsonBody(body), { signal, method: 'POST' }])
+        );
       },
 
       deleteEntityColumn: (type, attributeName) => {
-        return fetchRawls(`${root}/entities/${type}?attributeNames=${attributeName}`, _.mergeAll([authOpts(), { signal, method: 'DELETE' }]));
+        return fetchRawls(
+          `${root}/entities/${type}?attributeNames=${attributeName}`,
+          _.mergeAll([authOpts(), { signal, method: 'DELETE' }])
+        );
       },
 
       renameEntityColumn: (type, attributeName, newAttributeName) => {
@@ -424,19 +411,29 @@ const Workspaces = (signal) => ({
       },
 
       upsertEntities: (entityUpdates) => {
-        return fetchRawls(`${root}/entities/batchUpsert`, _.mergeAll([authOpts(), jsonBody(entityUpdates), { signal, method: 'POST' }]));
+        return fetchRawls(
+          `${root}/entities/batchUpsert`,
+          _.mergeAll([authOpts(), jsonBody(entityUpdates), { signal, method: 'POST' }])
+        );
       },
 
       paginatedEntitiesOfType: async (type, parameters) => {
-        const res = await fetchRawls(`${root}/entityQuery/${type}?${qs.stringify(parameters)}`, _.merge(authOpts(), { signal }));
+        const res = await fetchRawls(
+          `${root}/entityQuery/${type}?${qs.stringify(parameters)}`,
+          _.merge(authOpts(), { signal })
+        );
         return res.json();
       },
 
       deleteEntities: (entities) => {
-        return fetchRawls(`${root}/entities/delete`, _.mergeAll([authOpts(), jsonBody(entities), { signal, method: 'POST' }]));
+        return fetchRawls(
+          `${root}/entities/delete`,
+          _.mergeAll([authOpts(), jsonBody(entities), { signal, method: 'POST' }])
+        );
       },
 
-      getEntities: (entityType) => fetchRawls(`${root}/entities/${entityType}`, _.merge(authOpts(), { signal })).then((r) => r.json()),
+      getEntities: (entityType) =>
+        fetchRawls(`${root}/entities/${entityType}`, _.merge(authOpts(), { signal })).then((r) => r.json()),
 
       getEntitiesTsv: (entityType) =>
         fetchOrchestration(
@@ -473,7 +470,10 @@ const Workspaces = (signal) => ({
           return { name, entityType, operations: attributesUpdateOps(attributes) };
         }, payload);
 
-        return fetchRawls(`${root}/entities/batchUpsert`, _.mergeAll([authOpts(), jsonBody(body), { signal, method: 'POST' }]));
+        return fetchRawls(
+          `${root}/entities/batchUpsert`,
+          _.mergeAll([authOpts(), jsonBody(body), { signal, method: 'POST' }])
+        );
       },
 
       importEntitiesFile: (file, { deleteEmptyValues = false } = {}) => {
@@ -519,11 +519,14 @@ const Workspaces = (signal) => ({
       },
 
       listImportJobs: async (isRunning) => {
-        const res = await fetchOrchestration(`api/${root}/importJob?running_only=${isRunning}`, _.merge(authOpts(), { signal }));
+        const res = await fetchOrchestration(
+          `api/${root}/importJob?running_only=${isRunning}`,
+          _.merge(authOpts(), { signal })
+        );
         return res.json();
       },
 
-      importSnapshot: async (snapshotId, name, description) => {
+      importSnapshot: async (snapshotId, name, description?: any) => {
         const res = await fetchRawls(
           `${root}/snapshots/v2`,
           _.mergeAll([authOpts(), jsonBody({ snapshotId, name, description }), { signal, method: 'POST' }])
@@ -534,7 +537,10 @@ const Workspaces = (signal) => ({
       importAttributes: (file) => {
         const formData = new FormData();
         formData.set('attributes', file);
-        return fetchOrchestration(`api/${root}/importAttributesTSV`, _.merge(authOpts(), { body: formData, signal, method: 'POST' }));
+        return fetchOrchestration(
+          `api/${root}/importAttributesTSV`,
+          _.merge(authOpts(), { body: formData, signal, method: 'POST' })
+        );
       },
 
       exportAttributes: async () => {
@@ -543,12 +549,18 @@ const Workspaces = (signal) => ({
       },
 
       storageCostEstimate: async () => {
-        const res = await fetchOrchestration(`api/workspaces/${namespace}/${name}/storageCostEstimate`, _.merge(authOpts(), { signal }));
+        const res = await fetchOrchestration(
+          `api/workspaces/${namespace}/${name}/storageCostEstimate`,
+          _.merge(authOpts(), { signal })
+        );
         return res.json();
       },
 
       getTags: async () => {
-        const res = await fetchOrchestration(`api/workspaces/${namespace}/${name}/tags`, _.merge(authOpts(), { signal, method: 'GET' }));
+        const res = await fetchOrchestration(
+          `api/workspaces/${namespace}/${name}/tags`,
+          _.merge(authOpts(), { signal, method: 'GET' })
+        );
         return res.json();
       },
 
@@ -581,176 +593,4 @@ const Workspaces = (signal) => ({
   },
 });
 
-const FirecloudBucket = (signal) => ({
-  getServiceAlerts: async () => {
-    const res = await fetchOk(`${getConfig().firecloudBucketRoot}/alerts.json`, { signal });
-    return res.json();
-  },
-
-  getFeaturedWorkspaces: async () => {
-    const res = await fetchOk(`${getConfig().firecloudBucketRoot}/featured-workspaces.json`, { signal });
-    return res.json();
-  },
-
-  getShowcaseWorkspaces: async () => {
-    const res = await fetchOk(`${getConfig().firecloudBucketRoot}/showcase.json`, { signal });
-    return res.json();
-  },
-
-  getFeaturedMethods: async () => {
-    const res = await fetchOk(`${getConfig().firecloudBucketRoot}/featured-methods.json`, { signal });
-    return res.json();
-  },
-
-  getTemplateWorkspaces: async () => {
-    const res = await fetchOk(`${getConfig().firecloudBucketRoot}/template-workspaces.json`, { signal });
-    return res.json();
-  },
-
-  getBadVersions: async () => {
-    const res = await fetchOk(`${getConfig().firecloudBucketRoot}/bad-versions.txt`, { signal });
-    return res.text();
-  },
-});
-
-const Methods = (signal) => ({
-  list: async (params) => {
-    const res = await fetchAgora(`methods?${qs.stringify(params)}`, _.merge(authOpts(), { signal }));
-    return res.json();
-  },
-
-  definitions: async () => {
-    const res = await fetchAgora('methods/definitions', _.merge(authOpts(), { signal }));
-    return res.json();
-  },
-
-  configInputsOutputs: async (loadedConfig) => {
-    const res = await fetchRawls(
-      'methodconfigs/inputsOutputs',
-      _.mergeAll([authOpts(), jsonBody(loadedConfig.methodRepoMethod), { signal, method: 'POST' }])
-    );
-    return res.json();
-  },
-
-  template: async (modifiedConfigMethod) => {
-    const res = await fetchRawls('methodconfigs/template', _.mergeAll([authOpts(), jsonBody(modifiedConfigMethod), { signal, method: 'POST' }]));
-    return res.json();
-  },
-
-  method: (namespace, name, snapshotId) => {
-    const root = `methods/${namespace}/${name}/${snapshotId}`;
-
-    return {
-      get: async () => {
-        const res = await fetchAgora(root, _.merge(authOpts(), { signal }));
-        return res.json();
-      },
-
-      configs: async () => {
-        const res = await fetchAgora(`${root}/configurations`, _.merge(authOpts(), { signal }));
-        return res.json();
-      },
-
-      allConfigs: async () => {
-        const res = await fetchAgora(`methods/${namespace}/${name}/configurations`, _.merge(authOpts(), { signal }));
-        return res.json();
-      },
-
-      toWorkspace: async (workspace, config = {}) => {
-        const res = await fetchRawls(
-          `workspaces/${workspace.namespace}/${workspace.name}/methodconfigs`,
-          _.mergeAll([
-            authOpts(),
-            jsonBody(
-              _.merge(
-                {
-                  methodRepoMethod: {
-                    methodUri: `agora://${namespace}/${name}/${snapshotId}`,
-                  },
-                  name,
-                  namespace,
-                  rootEntityType: '',
-                  prerequisites: {},
-                  inputs: {},
-                  outputs: {},
-                  methodConfigVersion: 1,
-                  deleted: false,
-                },
-                config.payloadObject
-              )
-            ),
-            { signal, method: 'POST' },
-          ])
-        );
-        return res.json();
-      },
-    };
-  },
-});
-
-const Submissions = (signal) => ({
-  queueStatus: async () => {
-    const res = await fetchRawls('submissions/queueStatus', _.merge(authOpts(), { signal }));
-    return res.json();
-  },
-});
-
-const DrsUriResolver = (signal) => ({
-  // DRSHub now gets a signed URL instead of Martha
-  getSignedUrl: async ({ bucket, object, dataObjectUri, googleProject }) => {
-    const res = await fetchDrsHub(
-      '/api/v4/gcs/getSignedUrl',
-      _.mergeAll([jsonBody({ bucket, object, dataObjectUri, googleProject }), authOpts(), appIdentifier, { signal, method: 'POST' }])
-    );
-    return res.json();
-  },
-
-  getDataObjectMetadata: async (url, fields) => {
-    const res = await fetchDrsHub(
-      '/api/v4/drs/resolve',
-      _.mergeAll([jsonBody({ url, fields }), authOpts(), appIdentifier, { signal, method: 'POST' }])
-    );
-    return res.json();
-  },
-});
-
-const Surveys = (signal) => ({
-  submitForm: (formId, data) => fetchGoogleForms(`${formId}/formResponse?${qs.stringify(data)}`, { signal }),
-});
-
-export const Ajax = (signal) => {
-  return {
-    Apps: Apps(signal),
-    AzureStorage: AzureStorage(signal),
-    Billing: Billing(signal),
-    Buckets: GoogleStorage(signal),
-    Catalog: Catalog(signal),
-    Cbas: Cbas(signal),
-    CromIAM: CromIAM(signal),
-    CromwellApp: CromwellApp(signal),
-    DataRepo: DataRepo(signal),
-    Disks: Disks(signal),
-    Dockstore: Dockstore(signal),
-    DrsUriResolver: DrsUriResolver(signal),
-    ExternalCredentials: ExternalCredentials(signal),
-    FirecloudBucket: FirecloudBucket(signal),
-    Groups: Groups(signal),
-    Methods: Methods(signal),
-    Metrics: Metrics(signal),
-    OAuth2: OAuth2(signal),
-    Runtimes: Runtimes(signal),
-    SamResources: SamResources(signal),
-    Submissions: Submissions(signal),
-    Support: Support(signal),
-    Surveys: Surveys(signal),
-    TermsOfService: TermsOfService(signal),
-    User: User(signal),
-    WorkflowScript: WorkflowScript(signal),
-    WorkspaceData: WorkspaceData(signal),
-    WorkspaceManagerResources: WorkspaceManagerResources(signal),
-    Workspaces: Workspaces(signal),
-  };
-};
-
-// Exposing Ajax for use by integration tests (and debugging, or whatever)
-window.Ajax = Ajax;
+export type WorkspacesAjaxContract = ReturnType<typeof Workspaces>;
